Add header banner and style it
This is a *lot* of CSS to make this happen. We maybe don't need quite
this much at this stage, but we'll probably need it before long.

It doesn't include any reset CSS, so the visual isn't actually great as
the body still has a margin.

Mostly, this is just enough to demonstrate that the SCSS compiling stuff
is all set up and working properly.