Add deliver config for deployment to S3
If you want to serve HTML files without the extension from S3 you need
the files to have a `text/html` content type and to not have the
extension.

The command to generate the files to deploy first uses Jekyll to build
them, and then renames all `.html` files to drop the extension.