Add s3cmd config for deployment
The access key and secret key are omitted. These should be set as
environment variables, eg:

    AWS_ACCESS_KEY=hello AWS_SECRET_KEY=world deliver

An interesting part of this config is default_mime_type. Because we're
uploading HTML files without an extension, s3cmd doesn't know what mime
type to use. We default it to `text/html` so that S3 will serve the
files properly and the browser doesn't try to download them.