Show single-page conditions on index page
The filter used in the index page meant we only showed pages with a
`nav_order` of 1. That made sure we only got the first page for a
condition. It also made sure we didn't get the index page or help pages,
because they didn't have a `nav_order` at all.

Now that we have single-page conditions, setting a `nav_order` on them
seems wrong. Instead of using the `where` filter we use two
conditionals:

  1. Is `condition` set?
  2. Is `nav_order` empty or 1?

That means we get all single-page conditions, and the first page of all
multi-page conditions.