Remove link to diabetes.co.uk
As Said rightly points out here:
http://forum.digital.nhs.uk/t/feedback-on-type-2-diabetes-condition-pages/44/4

We are incorrectly attributing a `diabates.co.uk` page to Diabetes UK
(`diabetes.org.uk`).

For now I'm removing this error by deleting the link.

Separately, we should decide how or whether to reinstate this link (correctly
attributed). As Said points out they are a commercial organisation with
their own agenda, and we need a position on how/whether we link to them.